requirements.txt
Pillow>=10.0.0
PyQt5>=5.15.0
requests>=2.28.0
python-dotenv>=1.0.0
markdown>=3.5.0
pyinstaller>=6.0.0
